EventWatcher progress

+++ include/GC2D/GCConsole2.hpp
@@ -42,7 +42,7 @@ public:
 	void startAppearTelop(bool);
 	void startDisappearTelop();
 	void startDisappearTimer();
-	void startAppearTimer(int, long);
+	void startAppearTimer(int, s32);
 	void startInsertTimer();
 	void startAppearJetBalloon(int, int);
 	void startInsertJetBalloon();
@@ -58,7 +58,7 @@ public:
 	void startAppearMario(bool);
 	void processMoveNozzle();
 	void changeNum(TBlendPane*, int, int);
-	void setTimer(long);
+	void setTimer(s32);
 	void startMoveTimer(int);
 	void stopMoveTimer();
 	int getFinishedTime();
